isky-building-mservice/requirements.txt
aniso8601==9.0.1
appwrite==0.10.0
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
dependency-injector==4.39.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
flask-openapi3==1.1.4
flask-restx==0.5.1
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
mistune==2.0.2
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
